Вынес иконки навбара по своим папкам и структурировал их
