+++ src/shared/ui/icons/index.ts
@@ -1,0 +1,4 @@
+export { HeartIcon } from "./HeartIcon";
+export { SearchIcon } from "./SearchIcon";
+export { LogoIcon } from "./LogoIcon";
+export { MenuIcon } from "./MenuIcon";
