--- src/shared/ui/navbar/index.ts
@@ -1,1 +1,0 @@
-export { NavbarShowcase } from "./Navbar"
