+++ src/widgets/navbar/index.ts
@@ -1,0 +1,1 @@
+export { NavbarShowcase } from "./ui/Navbar"
